add: example of how we can document the api with dokka and kdoc and GitHub flavored markdown (gfm)
This can be done with HTML format also if preferred.

The action runs on publish and updates doc folder to contain the latest release documentation updates.

Also changed all workflows with gradle cache setup, to use action `setup-java` built-in cache instead of separate action.

Please add documentation before approve/merge, as this is lacking for the library due to some developer `laziness` :smile: