build: set kotlin compiler targets for user compatibility
This re-introduces compatibility for consuming users dependant on earlier
versions of the Kotlin compiler, while still allowing usage of latest
Kotlin toolchains within the project itself.

We've set the minimum supported version to 1.9.0 as of this commit,
which should revert the breaking change introduced with version
[2.1.1](https://github.com/navikt/mock-oauth2-server/releases/tag/2.1.11),
where the compiler minor version was incremented.

Huge thanks to this massively useful blog post:
https://blog.alllex.me/posts/2023-11-03-liberal-library-tooling/